requirements.txt
flask==2.1.2
python-dotenv==0.20.0
langchain==0.0.184
openai==0.27.0
werkzeug==2.3.4